Commit with new texts

+++ src/en/com/dio/bootcamp/Content.java
@@ -2,12 +2,12 @@ package en.com.dio.bootcamp;
 
 public abstract class Content {
 
-    protected static final double BASE_XP = 10.0;
+    protected static final int BASE_XP = 10;
 
     private String name;
     private String description;
 
-    public abstract double determineXp();
+    public abstract int determineXp();
 
     public static double getBaseXp() {
         return BASE_XP;
@@ -28,5 +28,6 @@ public abstract class Content {
     public void setDescription(String description) {
         this.description = description;
     }
+
     
 }
